move controller and tests out of stories
